+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Set environment variables in a seed (seed toml files only for now)
+
 ## [0.2.0] - 2025-09-27
 
 ### Added