+++ CHANGELOG.md
@@ -8,24 +8,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
 - Set environment variables in a seed (seed toml files only for now)
 
 ### Changed
+
 - Make desktop entry category a configurable option
-- Show app category when listing apps 
+- Show app category when listing apps
 - Make default proton runtime config option not mandatory
+- Add environment variable information to seed information
 
 ## [0.2.0] - 2025-09-27
 
 ### Added
+
 - Ability to generate menus for individual app entries
 - Ability to use local `umu-run`
 - Ability to use custom runtime folder from Flatpak Steam
 
 ### Fixed
+
 - Errors while generating menus if required directories does not exist by creating the required directories first
 
 ### Changed
+
 - BREAKING CHANGE: Move in `generate-menu` to an `app` subcommand, since it made more sense there
 - BREAKING CHANGE: Shorten app-related flags in `flora` CLI
   - `default-application-name` -> `app-name`
@@ -38,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2025-09-10
 
 ### Added
+
 - Manage Wine and Proton setups, configured in a `seed`. Each `seed` can have separate prefixes and runtimes.
 - Define application entries to be launched from a `seed`.
 - Generate application entries from Start Menu shortcuts.