+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use custom Wine and Proton runtimes for a `seed`.
 - Transparent configuration, everything is laid out in `toml` files
 
-[unreleased]: https://github.com/Damillora/flora/compare/v0.2.0...HEAD
+[unreleased]: https://github.com/Damillora/flora/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/Damillora/flora/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/Damillora/flora/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/Damillora/flora/releases/tag/v0.1.0